+++ packages/react-styled-ui/src/Scrollbar/utils/getInnerHeight.js
@@ -1,0 +1,7 @@
+const getInnerHeight = (el) => {
+  const { clientHeight } = el;
+  const { paddingTop, paddingBottom } = getComputedStyle(el);
+  return clientHeight - parseFloat(paddingTop) - parseFloat(paddingBottom);
+};
+
+export default getInnerHeight;
