+++ packages/react-styled-ui/src/Scrollbar/utils/getInnerWidth.js
@@ -1,0 +1,7 @@
+const getInnerWidth = (el) => {
+  const { clientWidth } = el;
+  const { paddingLeft, paddingRight } = getComputedStyle(el);
+  return clientWidth - parseFloat(paddingLeft) - parseFloat(paddingRight);
+};
+
+export default getInnerWidth;
