+++ packages/styled-ui-docs/pages/scrollbar.mdx
@@ -11,8 +11,8 @@ import {
 ## Usage
 
 ```jsx
-<Scrollbar height={120}>
-  <Lorem count={12} />
+<Scrollbar height={120} width={600}>
+  <Lorem count={12} style={{ whiteSpace: 'nowrap' }}  />
 </Scrollbar>
 ```
